NFM: buildsys: Do not use pkgconfig for finding libusb
The idea behing this change is to allow the libusb include directory and
the libusb binary to be manually specified if pkgconfig cannot find it.

This is useless!!! MSVC does not support C99, and thus there is no reason
why pkgconfig could get in the way with other compilers. Until this change
can be justified, it will NOT be merged.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>