libqiprog: Remove set_address() from the API (FTBFS)
set_address() is an artifact of the USB protocol, and after discussing it
with Peter, we agreed to remove it from the API. set_address() will
continue to be part of the usb driver, albeit hidden from the rest of the
world.

set_address() is also removed from the core, but remains as an optional
component of the private driver struct.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>