usb_master: Fix uninitialized variable when resubmitting transfers
In the async callback, we are assuming that 'ret' is initialized to
QIPROG_SUCCESS before checking for errors, however it was never
initialized. This caused the transfer to not be resubmitted, without
any error message. The transfer was then removed from the list if
active transfers, and the loop returned without having transmitted the
complete dataset.
Initialize 'ret' to QIPROG_SUCCESS before any error checking.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>